Moved convenience method from model
